wip: use sol bindings
