+++ crates/tempo-zone/Cargo.toml
@@ -44,6 +44,7 @@ reth-transaction-pool.workspace = true
 
 # alloy
 alloy-consensus.workspace = true
+alloy-contract.workspace = true
 alloy-primitives.workspace = true
 alloy-provider.workspace = true
 alloy-rlp.workspace = true
